% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{adj_mat_to_long}
\alias{adj_mat_to_long}
\title{Converts an adjacency matrix to a long format dataframe}
\usage{
adj_mat_to_long(W, include_zeros = FALSE)
}
\arguments{
\item{W}{An adjacency matrix}
}
\value{
A dataframe with columns `from`, `to`, and `value`
}
\description{
Converts an adjacency matrix to a long format dataframe
}
