% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_expm.R
\name{h_expm}
\alias{h_expm}
\alias{h_expm_grad}
\title{Implements the matrix exponential acyclicity regularizer from DAGs with no TEARS (Zheng et al 2018)}
\usage{
h_expm(W, include_grad = FALSE)

h_expm_grad(W)
}
\arguments{
\item{W}{weighted adjacency matrix}

\item{include_grad}{whether to include the gradient of the regularizer}
}
\value{
either a list if gradient is included or a scalar
}
\description{
Implements the matrix exponential acyclicity regularizer from DAGs with no TEARS (Zheng et al 2018)
}
\examples{
 X <- matrix(0, nrow = 3, ncol = 3)
X_lower <- lower.tri(X)

# DAG
h_expm(lower.tri(X))

# Not a DAG
X2 <- X_lower
X2[1, 2] <- 1
h_expm(X2)
}
