% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{is_dag}
\alias{is_dag}
\title{Checks if matrix is a directed acyclic graph}
\usage{
is_dag(W, threshold = 1e-05)
}
\arguments{
\item{W}{A matrix that represents a directed graph}

\item{threshold}{A threshold to determine if a value is zero}
}
\value{
A boolean indicating if the matrix is a directed acyclic graph
}
\description{
Checks if matrix is a directed acyclic graph
}
\examples{
d <- 100
B <- matrix(0, d, d)
idx <- sample(1:d)
B[lower.tri(B)] <- runif(d*(d - 1) / 2, min = 1, max = 100)
is_dag(B[idx, idx]) # DAG

B[1,10] <- 1
is_dag(B[idx, idx]) # Not a DAG
}
