% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_linear_sem.R
\name{sim_linear_sem}
\alias{sim_linear_sem}
\title{Simulate from linear SEM}
\usage{
sim_linear_sem(W, n = 1, Sigma = diag(ncol(W)))
}
\arguments{
\item{W}{Adjacency matrix representing SEM (d x d)}

\item{n}{Number of samples}

\item{Sigma}{Covariance matrix of noise}
}
\value{
Matrix of samples (n x d)
}
\description{
Simulate from the model X = X %*% W + eps
If we pass in upper triangular then we take transpose
X ~ MVN(0, (I - W)^-T %*% Sigma %*% (I - W)^-1)
Where eps ~ MVN(0, Sigma)
}
\examples{
B <- matrix(
c(0, 3, 0, 3,
  0, 0, 0, 5,
  0, 0, 0, 2,
  0, 0, 0, 0),
nrow = 4, ncol = 4, byrow = TRUE)
d <- ncol(B)
X <- sim_linear_sem(B, n = 500, Sigma = 1 * diag(ncol(B)))
}
