% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_utils.R
\name{spectral_radius}
\alias{spectral_radius}
\title{Computes the spectral radius (the largest absolute eigenvalue) of a matrix}
\usage{
spectral_radius(W)
}
\arguments{
\item{W}{A matrix}
}
\value{
The spectral radius of the matrix
}
\description{
Computes the spectral radius (the largest absolute eigenvalue) of a matrix
}
\examples{
B <- matrix(c(0, 0.6, -0.5, 0), nrow = 2, byrow = TRUE)
spectral_radius(B)
}
