% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_matrix.R
\name{threshold_W}
\alias{threshold_W}
\title{Threshold a matrix by setting the variables less than threshold to 0}
\usage{
threshold_W(W, threshold = 0.1)
}
\arguments{
\item{W}{A matrix}

\item{threshold}{A threshold value}
}
\value{
W with values less than threshold set to 0
}
\description{
Threshold a matrix by setting the variables less than threshold to 0
}
\examples{
W <- matrix(c(0, 0.6, -0.5, 1e-3), nrow = 2, byrow = TRUE)
threshold_W(W, threshold = 0.1)
}
