% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_to_direct.R
\name{total_to_direct}
\alias{total_to_direct}
\alias{direct_to_total}
\title{Converts total effects to direct effects and visa verse}
\usage{
total_to_direct(W, restrict_dag = TRUE, enforce_spectral_radius = TRUE)

direct_to_total(W, restrict_dag = TRUE)
}
\arguments{
\item{W}{A matrix of total/direct linear effects}

\item{restrict_dag}{A boolean indicating if the matrix must be a directed acyclic graph}

\item{enforce_spectral_radius}{A boolean indicating if the spectral radius of W_{dir} must be less than 1. Generally not recommended to disable}
}
\value{
A matrix of direct effects
}
\description{
Assuming a matrix of total linear effects W we can convert this to a matrix of direct effects W_{dir}
This assume that W_{tot} = W_{dir} + W_{dir}^2 + W_{dir}^3 + ...
We can solve for W_{dir} = I - (I + W_{tot})^{-1} assume that the spectral radius (largest absolute eigenvalue) of W_{dir} < 1
This condition is the same that abs(eigenvalue(W_{tot}) / (1 + eigenvalue(W_{tot}))) < 1
}
\examples{
B <- matrix(c(0, 0.6, -0.5, 0), nrow = 2, byrow = TRUE)
total_to_direct(B,  restrict_dag = F)
all(zapsmall(direct_to_total(total_to_direct(B, restrict_dag = F), restrict_dag = F)) == B)

B2 <- matrix(c(0,0,0,
             1,0,0,
             1,1,0), nrow = 3, byrow = TRUE)
total_to_direct(B2)
}
